refactor: use custom type static struct to parse html on startup
the motivation behind this change is that parsing an html file at every request is not ideal. also, if the html is corrupted, we will only see the error when we try to load the page. for that reason, it would be better if we parsed the html on startup, then execute the file at every request