feat: require user via middleware
section 17 - requiring a user via middleware

Some endpoints are going to require that a user is logged in to access
them. We need a middleware that helps with getting the user from the
request then setting the user before proceeding. If no user is found,
our server will respond with a redirect to the sign in page.