use template for emails
not the best code organisation, but just to get the idea. had a hard time making it work, kept getting the error template is incomplete or empty template, turned out to be the template.New(name) name that was wrong, it has to be the name of the file, not just a name. see docs and link in the code for reference