fix: potential bug when reading template under subfolder in filesystem
when trying to read the templates/galleries/new.gothml, we got an error
'template: galleries/new.gohtml is an incomplete or empty template'

this is caused becase the within ParseFS function we pass the full
pathname to the builtin template parser. under the hood, the parser uses
`filepath.Base() or path.Base()` to get the filename, and that is the
new.gohtml only.

our fix works, but having multiple files with the same name - ie
new.gothml - might cause trouble down the road.